requirements.txt
Flask==2.3.3
pdfplumber==0.9.0
python-docx==0.8.11
sentence-transformers==2.2.2
spacy==3.7.2
torch>=1.9.0
numpy==1.24.3
scikit-learn==1.3.0